game-controller/requirements.txt
paho-mqtt==2.1.0
colorama==0.4.6